Admin alatti kontrollerek átalakítása
