+++ src/BaseController.php
@@ -30,4 +30,12 @@ abstract class BaseController
     {
         (new ErrorController($this->entityManager))->index();
     }
+
+    protected function checkAdminAuth(): void
+    {
+        if (!isset($_SESSION['admin_id'])) {
+            header('Location: /admin/login');
+            exit;
+        }
+    }
 }
