+++ src/Templates/Admin/admin_header.php
@@ -36,7 +36,7 @@ if (isset($statusCode)) {
                 </ul>
                 <ul class="navbar-nav ms-auto">
                     <li class="nav-item">
-                        <a class="nav-link" href="/admin/logout">Kijelentkezés</a>
+                        <a class="nav-link" href="/admin/login/logout">Kijelentkezés</a>
                     </li>
                 </ul>
             </div>
